src/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
async-tio==1.3.2
attrs==24.2.0
azure-cognitiveservices-speech==1.41.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mysql-connector-python==9.1.0
openai==1.52.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.1.1
regex==2024.9.11
requests==2.32.3
rfc3986==2.0.0
rich==13.9.3
sniffio==1.3.1
socksio==1.0.0
tiktoken==0.8.0
tls-client==1.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.16.0
